[client_utils] move client_utils headers to include/ directory
So we can install them to a proper location
