+++ client_utils/include/hawktracer/client_utils.hpp
@@ -1,0 +1,8 @@
+#ifndef HAWKTRACER_CLIENT_UTILS_HPP
+#define HAWKTRACER_CLIENT_UTILS_HPP
+
+#include <hawktracer/client_utils/command_line_parser.hpp>
+#include <hawktracer/client_utils/stream_factory.hpp>
+#include <hawktracer/client_utils/tcp_client_stream.hpp>
+
+#endif // HAWKTRACER_CLIENT_UTILS_HPP
